managed the .gvfs_meta file by DefaultRawFileSystem

+++ clients/filesystem-fuse/src/default_raw_filesystem.rs
@@ -18,10 +18,11 @@
  */
 use crate::config::AppConfig;
 use crate::filesystem::{
-    FileStat, FileSystemContext, PathFileSystem, RawFileSystem, Result, INITIAL_FILE_ID,
-    ROOT_DIR_FILE_ID, ROOT_DIR_PARENT_FILE_ID, ROOT_DIR_PATH,
+    FileStat, FileSystemContext, PathFileSystem, RawFileSystem, Result, FS_META_FILE_ID,
+    FS_META_FILE_NAME, FS_META_FILE_PATH, INITIAL_FILE_ID, ROOT_DIR_FILE_ID,
+    ROOT_DIR_PARENT_FILE_ID, ROOT_DIR_PATH,
 };
-use crate::opened_file::{FileHandle, OpenFileFlags};
+use crate::opened_file::{FileHandle, OpenFileFlags, OpenedFile};
 use crate::opened_file_manager::OpenedFileManager;
 use async_trait::async_trait;
 use bytes::Bytes;
@@ -133,6 +134,19 @@ impl<T: PathFileSystem> DefaultRawFileSystem<T> {
         let mut file_manager = self.file_entry_manager.write().await;
         file_manager.insert(parent_file_id, file_id, path);
     }
+
+    fn meta_file_stat(&self) -> FileStat {
+        let mut meta_file_stat =
+            FileStat::new_file_filestat_with_path(Path::new(FS_META_FILE_PATH), 0);
+        meta_file_stat.set_file_id(ROOT_DIR_FILE_ID, FS_META_FILE_ID);
+        meta_file_stat
+    }
+
+    fn root_file_stat(&self) -> FileStat {
+        let mut root_file_stat = FileStat::new_dir_filestat_with_path(Path::new(ROOT_DIR_PATH));
+        root_file_stat.set_file_id(ROOT_DIR_PARENT_FILE_ID, ROOT_DIR_FILE_ID);
+        root_file_stat
+    }
 }
 
 #[async_trait]
@@ -143,6 +157,13 @@ impl<T: PathFileSystem> RawFileSystem for DefaultRawFileSystem<T> {
             ROOT_DIR_PARENT_FILE_ID,
             ROOT_DIR_FILE_ID,
             Path::new(ROOT_DIR_PATH),
+        )
+        .await;
+
+        self.insert_file_entry_locked(
+            ROOT_DIR_FILE_ID,
+            FS_META_FILE_ID,
+            Path::new(FS_META_FILE_PATH),
         )
         .await;
         self.fs.init().await
@@ -169,6 +190,13 @@ impl<T: PathFileSystem> RawFileSystem for DefaultRawFileSystem<T> {
     }
 
     async fn stat(&self, file_id: u64) -> Result<FileStat> {
+        if file_id == ROOT_DIR_FILE_ID {
+            return Ok(self.root_file_stat());
+        }
+        if file_id == FS_META_FILE_ID {
+            return Ok(self.meta_file_stat());
+        }
+
         let file_entry = self.get_file_entry(file_id).await?;
         let mut file_stat = self.fs.stat(&file_entry.path).await?;
         file_stat.set_file_id(file_entry.parent_file_id, file_entry.file_id);
@@ -176,8 +204,11 @@ impl<T: PathFileSystem> RawFileSystem for DefaultRawFileSystem<T> {
     }
 
     async fn lookup(&self, parent_file_id: u64, name: &OsStr) -> Result<FileStat> {
-        let parent_file_entry = self.get_file_entry(parent_file_id).await?;
+        if parent_file_id == ROOT_DIR_FILE_ID && name == OsStr::new(FS_META_FILE_NAME) {
+            return Ok(self.meta_file_stat());
+        }
 
+        let parent_file_entry = self.get_file_entry(parent_file_id).await?;
         let path = parent_file_entry.path.join(name);
         let mut file_stat = self.fs.stat(&path).await?;
         // fill the file id to file stat
@@ -193,10 +224,21 @@ impl<T: PathFileSystem> RawFileSystem for DefaultRawFileSystem<T> {
         for file_stat in child_filestats.iter_mut() {
             self.resolve_file_id_to_filestat(file_stat, file_id).await;
         }
+
+        if file_id == ROOT_DIR_FILE_ID {
+            child_filestats.push(self.meta_file_stat());
+        }
         Ok(child_filestats)
     }
 
     async fn open_file(&self, file_id: u64, flags: u32) -> Result<FileHandle> {
+        if file_id == FS_META_FILE_ID {
+            let meta_file = OpenedFile::new(self.meta_file_stat());
+            let resutl = self.opened_file_manager.put(meta_file);
+            let file = resutl.lock().await;
+            return Ok(file.file_handle());
+        }
+
         self.open_file_internal(file_id, flags, FileType::RegularFile)
             .await
     }
@@ -212,6 +254,10 @@ impl<T: PathFileSystem> RawFileSystem for DefaultRawFileSystem<T> {
         name: &OsStr,
         flags: u32,
     ) -> Result<FileHandle> {
+        if parent_file_id == ROOT_DIR_FILE_ID && name == OsStr::new(FS_META_FILE_NAME) {
+            return Err(Errno::from(libc::EEXIST));
+        }
+
         let parent_file_entry = self.get_file_entry(parent_file_id).await?;
         let mut file_without_id = self
             .fs
@@ -248,11 +294,19 @@ impl<T: PathFileSystem> RawFileSystem for DefaultRawFileSystem<T> {
     }
 
     async fn set_attr(&self, file_id: u64, file_stat: &FileStat) -> Result<()> {
+        if file_id == ROOT_DIR_FILE_ID || file_id == FS_META_FILE_ID {
+            return Ok(());
+        }
+
         let file_entry = self.get_file_entry(file_id).await?;
         self.fs.set_attr(&file_entry.path, file_stat, true).await
     }
 
     async fn remove_file(&self, parent_file_id: u64, name: &OsStr) -> Result<()> {
+        if parent_file_id == ROOT_DIR_FILE_ID && name == OsStr::new(FS_META_FILE_NAME) {
+            return Err(Errno::from(libc::EPERM));
+        }
+
         let parent_file_entry = self.get_file_entry(parent_file_id).await?;
         let path = parent_file_entry.path.join(name);
         self.fs.remove_file(&path).await?;
@@ -281,7 +335,11 @@ impl<T: PathFileSystem> RawFileSystem for DefaultRawFileSystem<T> {
         file.close().await
     }
 
-    async fn read(&self, _file_id: u64, fh: u64, offset: u64, size: u32) -> Result<Bytes> {
+    async fn read(&self, file_id: u64, fh: u64, offset: u64, size: u32) -> Result<Bytes> {
+        if file_id == FS_META_FILE_ID {
+            return Ok(Bytes::new());
+        }
+
         let (data, file_stat) = {
             let opened_file = self
                 .opened_file_manager
@@ -298,7 +356,11 @@ impl<T: PathFileSystem> RawFileSystem for DefaultRawFileSystem<T> {
         data
     }
 
-    async fn write(&self, _file_id: u64, fh: u64, offset: u64, data: &[u8]) -> Result<u32> {
+    async fn write(&self, file_id: u64, fh: u64, offset: u64, data: &[u8]) -> Result<u32> {
+        if file_id == FS_META_FILE_ID {
+            return Err(Errno::from(libc::EPERM));
+        }
+
         let (len, file_stat) = {
             let opened_file = self
                 .opened_file_manager
